+++ tanjun/checks.py
@@ -36,11 +36,11 @@ from __future__ import annotations
 __all__: list[str] = [
     "CallbackReturnT",
     "CommandT",
-    "ApplicationOwnerCheck",
-    "nsfw_check",
-    "sfw_check",
-    "dm_check",
-    "guild_check",
+    "OwnerCheck",
+    "NsfwCheck",
+    "SfwCheck",
+    "DmCheck",
+    "GuildCheck",
     "PermissionCheck",
     "AuthorPermissionCheck",
     "OwnPermissionsCheck",
@@ -56,9 +56,6 @@ __all__: list[str] = [
 
 import abc
 import asyncio
-import dataclasses
-import datetime
-import time
 import typing
 from collections import abc as collections
 
@@ -67,6 +64,7 @@ import hikari
 from . import _backoff as backoff
 from . import errors
 from . import injecting
+from . import standard_dependencies
 from . import utilities
 
 if typing.TYPE_CHECKING:
@@ -112,158 +110,55 @@ class InjectableCheck(injecting.BaseInjectableCallback[bool]):
         raise errors.FailedCheck
 
 
-@dataclasses.dataclass(frozen=True)
-class _WrappedKwargs:
-    callback: tanjun_abc.CheckSig
-    _kwargs: dict[str, typing.Any]
-
-    def __call__(self, ctx: tanjun_abc.Context, /) -> tanjun_abc.MaybeAwaitableT[bool]:
-        return self.callback(ctx, **self._kwargs)
-
-    # This is delegated to the callback in-order to delegate set/list behaviour for this class to the callback.
-    def __eq__(self, other: typing.Any) -> bool:
-        return bool(self.callback == other)
-
-    # This is delegated to the callback in-order to delegate set/list behaviour for this class to the callback.
-    def __hash__(self) -> int:
-        return hash(self.callback)
-
-
-def _wrap_with_kwargs(
-    command: typing.Optional[CommandT],
-    callback: tanjun_abc.CheckSig,
-    /,
-    **kwargs: typing.Any,
-) -> CallbackReturnT[CommandT]:
+def _optional_kwargs(
+    command: typing.Optional[CommandT], check: tanjun_abc.CheckSig, /
+) -> typing.Union[CommandT, collections.Callable[[CommandT], CommandT]]:
     if command:
-        if kwargs:
-            return command.add_check(_WrappedKwargs(callback, kwargs))
+        return command.add_check(check)
 
-        return command.add_check(callback)
-
-    return lambda command_: command_.add_check(_WrappedKwargs(callback, kwargs))
+    return lambda c: c.add_check(check)
 
 
-def _handle_result(value: bool, error_message: typing.Optional[str], halt_execution: bool, /) -> bool:
-    if not value:
-        if error_message:
-            raise errors.CommandError(error_message)
-        if halt_execution:
-            raise errors.HaltExecution
+class _Check:
+    __slots__ = ("_error_message", "_halt_execution")
 
-    return value
+    def __init__(
+        self,
+        error_message: typing.Optional[str],
+        halt_execution: bool,
+    ) -> None:
+        self._error_message = error_message
+        self._halt_execution = halt_execution
+
+    def _handle_result(self, result: bool) -> bool:
+        if not result:
+            if self._error_message:
+                raise errors.CommandError(self._error_message)
+            if self._halt_execution:
+                raise errors.HaltExecution
+
+        return result
 
 
-class ApplicationOwnerCheck:
-    __slots__ = ("_application", "_error_message", "_expire", "_halt_execution", "_lock", "_owner_ids", "_time")
+class OwnerCheck(_Check):
+    __slots__ = ()
 
     def __init__(
         self,
         *,
         error_message: typing.Optional[str] = "Only bot owners can use this command",
-        expire_delta: datetime.timedelta = datetime.timedelta(minutes=5),
         halt_execution: bool = False,
-        owner_ids: typing.Optional[collections.Iterable[hikari.SnowflakeishOr[hikari.User]]] = None,
     ) -> None:
-        self._application: typing.Optional[hikari.Application] = None
-        self._error_message = error_message
-        self._expire = expire_delta.total_seconds()
-        self._halt_execution = halt_execution
-        self._lock = asyncio.Lock()
-        self._owner_ids = tuple(hikari.Snowflake(id_) for id_ in owner_ids) if owner_ids else ()
-        self._time = 0.0
+        super().__init__(error_message, halt_execution)
 
-    async def __call__(self, ctx: tanjun_abc.Context, /) -> bool:
-        return await self.check(ctx)
-
-    @property
-    def _is_expired(self) -> bool:
-        return time.perf_counter() - self._time >= self._expire
-
-    async def _try_fetch(self, rest: hikari.api.RESTClient, /) -> hikari.Application:  # type: ignore[return]
-        retry = backoff.Backoff()
-        async for _ in retry:
-            try:
-                self._application = await rest.fetch_application()  # TODO: or fetch authroization
-                return self._application
-
-            except (hikari.RateLimitedError, hikari.RateLimitTooLongError) as exc:
-                retry.set_next_backoff(exc.retry_after)
-
-            except hikari.InternalServerError:
-                continue
-
-    async def _get_application(self, ctx: tanjun_abc.Context, /) -> hikari.Application:
-        if self._application and not self._is_expired:
-            return self._application
-
-        async with self._lock:
-            if self._application and not self._is_expired:
-                return self._application
-
-            try:
-                application = await ctx.client.rest.fetch_application()
-                self._application = application
-
-            except (
-                hikari.RateLimitedError,
-                hikari.RateLimitTooLongError,
-                hikari.InternalServerError,
-            ):
-                # If we can't fetch this information straight away and don't have a stale state to go off then we
-                # have to retry before returning.
-                if not self._application:
-                    application = await asyncio.wait_for(self._try_fetch(ctx.client.rest), 10)
-
-                # Otherwise we create a task to ensure that we will still try to refresh the stored state in the future
-                # while returning the stale state to ensure that the command execution doesn't stall.
-                else:
-                    asyncio.create_task(asyncio.wait_for(self._try_fetch(ctx.client.rest), self._expire * 0.80))
-                    application = self._application
-
-            self._time = time.perf_counter()
-
-        return application
-
-    def close(self) -> None:
-        ...  # This is only left in to match up with the `open` method.
-
-    async def open(
+    async def __call__(
         self,
-        client: tanjun_abc.Client,
-        /,
-        *,
-        timeout: typing.Optional[datetime.timedelta] = datetime.timedelta(seconds=30),
-    ) -> None:
-        try:
-            await self.update(client.rest, timeout=timeout)
-
-        except asyncio.TimeoutError:
-            pass
-
-    async def check(self, ctx: tanjun_abc.Context, /) -> bool:
-        if ctx.author.id in self._owner_ids:
-            return True
-
-        application = await self._get_application(ctx)
-
-        if application.team:
-            result = ctx.author.id in application.team.members
-
-        else:
-            result = ctx.author.id == application.owner.id
-
-        return _handle_result(result, self._error_message, self._halt_execution)
-
-    async def update(
-        self,
-        rest: hikari.api.RESTClient,
-        /,
-        *,
-        timeout: typing.Optional[datetime.timedelta] = datetime.timedelta(seconds=30),
-    ) -> None:
-        self._time = time.perf_counter()
-        await asyncio.wait_for(self._try_fetch(rest), timeout.total_seconds() if timeout else None)
+        ctx: tanjun_abc.Context,
+        dependency: standard_dependencies.AbstractOwnerCheck = injecting.injected(
+            type=standard_dependencies.AbstractOwnerCheck
+        ),
+    ) -> bool:
+        return self._handle_result(await dependency.check_ownership(ctx.client, ctx.author))
 
 
 async def _get_is_nsfw(ctx: tanjun_abc.Context, /) -> bool:
@@ -278,48 +173,63 @@ async def _get_is_nsfw(ctx: tanjun_abc.Context, /) -> bool:
     return channel.is_nsfw or False if isinstance(channel, hikari.GuildChannel) else True
 
 
-async def nsfw_check(
-    ctx: tanjun_abc.Context,
-    /,
-    *,
-    error_message: typing.Optional[str] = "Command can only be used in NSFW channels",
-    halt_execution: bool = False,
-) -> bool:
+class NsfwCheck(_Check):
+    __slots__ = ()
 
-    return _handle_result(await _get_is_nsfw(ctx), error_message, halt_execution)
+    def __init__(
+        self,
+        error_message: typing.Optional[str] = "Command can only be used in NSFW channels",
+        halt_execution: bool = False,
+    ) -> None:
+        super().__init__(error_message, halt_execution)
 
-
-async def sfw_check(
-    ctx: tanjun_abc.Context,
-    /,
-    *,
-    error_message: typing.Optional[str] = "Command can only be used in SFW channels",
-    halt_execution: bool = False,
-) -> bool:
-    return _handle_result(not await _get_is_nsfw(ctx), error_message, halt_execution)
+    async def __call__(self, ctx: tanjun_abc.Context, /) -> bool:
+        return self._handle_result(await _get_is_nsfw(ctx))
 
 
-def dm_check(
-    ctx: tanjun_abc.Context,
-    /,
-    *,
-    error_message: typing.Optional[str] = "Command can only be used in DMs",
-    halt_execution: bool = False,
-) -> bool:
-    return _handle_result(ctx.guild_id is None, error_message, halt_execution)
+class SfwCheck(_Check):
+    __slots__ = ()
+
+    def __init__(
+        self,
+        error_message: typing.Optional[str] = "Command can only be used in SFW channels",
+        halt_execution: bool = False,
+    ) -> None:
+        super().__init__(error_message, halt_execution)
+
+    async def __call__(self, ctx: tanjun_abc.Context, /) -> bool:
+        return self._handle_result(not await _get_is_nsfw(ctx))
 
 
-def guild_check(
-    ctx: tanjun_abc.Context,
-    /,
-    *,
-    error_message: typing.Optional[str] = "Command can only be used in guild channels",
-    halt_execution: bool = False,
-) -> bool:
-    return _handle_result(ctx.guild_id is not None, error_message, halt_execution)
+class DmCheck(_Check):
+    __slots__ = ()
+
+    def __init__(
+        self,
+        error_message: typing.Optional[str] = "Command can only be used in DMs",
+        halt_execution: bool = False,
+    ) -> None:
+        super().__init__(error_message, halt_execution)
+
+    def __call__(self, ctx: tanjun_abc.Context, /) -> bool:
+        return self._handle_result(ctx.guild_id is None)
 
 
-class PermissionCheck(abc.ABC):
+class GuildCheck(_Check):
+    __slots__ = ()
+
+    def __init__(
+        self,
+        error_message: typing.Optional[str] = "Command can only be used in guild channels",
+        halt_execution: bool = False,
+    ) -> None:
+        super().__init__(error_message, halt_execution)
+
+    def __call__(self, ctx: tanjun_abc.Context, /) -> bool:
+        return self._handle_result(ctx.guild_id is not None)
+
+
+class PermissionCheck(_Check):
     __slots__ = ("_halt_execution", "_error_message", "permissions")
 
     def __init__(
@@ -330,13 +240,12 @@ class PermissionCheck(abc.ABC):
         error_message: typing.Optional[str],
         halt_execution: bool = False,
     ) -> None:
-        self._halt_execution = halt_execution
-        self._error_message = error_message
+        super().__init__(error_message, halt_execution)
         self.permissions = hikari.Permissions(permissions)
 
     async def __call__(self, ctx: tanjun_abc.Context, /) -> bool:
         result = (self.permissions & await self.get_permissions(ctx)) == self.permissions
-        return _handle_result(result, self._error_message, self._halt_execution)
+        return self._handle_result(result)
 
     @abc.abstractmethod
     async def get_permissions(self, ctx: tanjun_abc.Context, /) -> hikari.Permissions:
@@ -469,7 +378,7 @@ def with_dm_check(
     CallbackReturnT[CommandT]
         The command this check was added to.
     """
-    return _wrap_with_kwargs(command, dm_check, halt_execution=halt_execution, error_message=error_message)
+    return _optional_kwargs(command, DmCheck(halt_execution=halt_execution, error_message=error_message))
 
 
 @typing.overload
@@ -522,7 +431,7 @@ def with_guild_check(
     CallbackReturnT[CommandT]
         The command this check was added to.
     """
-    return _wrap_with_kwargs(command, guild_check, halt_execution=halt_execution, error_message=error_message)
+    return _optional_kwargs(command, GuildCheck(halt_execution=halt_execution, error_message=error_message))
 
 
 @typing.overload
@@ -575,7 +484,7 @@ def with_nsfw_check(
     CallbackReturnT[CommandT]
         The command this check was added to.
     """
-    return _wrap_with_kwargs(command, nsfw_check, halt_execution=halt_execution, error_message=error_message)
+    return _optional_kwargs(command, NsfwCheck(halt_execution=halt_execution, error_message=error_message))
 
 
 @typing.overload
@@ -630,7 +539,7 @@ def with_sfw_check(
     CallbackReturnT[CommandT]
         The command this check was added to.
     """
-    return _wrap_with_kwargs(command, sfw_check, halt_execution=halt_execution, error_message=error_message)
+    return _optional_kwargs(command, SfwCheck(halt_execution=halt_execution, error_message=error_message))
 
 
 @typing.overload
@@ -642,9 +551,7 @@ def with_owner_check(command: CommandT, /) -> CommandT:
 def with_owner_check(
     *,
     error_message: typing.Optional[str] = "Only bot owners can use this command",
-    expire_delta: datetime.timedelta = datetime.timedelta(minutes=5),
     halt_execution: bool = False,
-    owner_ids: typing.Optional[collections.Iterable[hikari.SnowflakeishOr[hikari.User]]] = None,
 ) -> collections.Callable[[CommandT], CommandT]:
     ...
 
@@ -654,9 +561,7 @@ def with_owner_check(
     /,
     *,
     error_message: typing.Optional[str] = "Only bot owners can use this command",
-    expire_delta: datetime.timedelta = datetime.timedelta(minutes=5),
     halt_execution: bool = False,
-    owner_ids: typing.Optional[collections.Iterable[hikari.SnowflakeishOr[hikari.User]]] = None,
 ) -> CallbackReturnT[CommandT]:
     """Only let a command run if it's being triggered by one of the bot's owners.
 
@@ -672,24 +577,14 @@ def with_owner_check(
 
         Defaults to "Only bot owners can use this command" and setting this to `None`
         will disable the error message allowing the command search to continue.
-    expire_delta: datetime.timedelta
-        How long cached application owner data should be cached for.
-
-        Defaults to 5 minutes.
     halt_execution : bool
         Whether this check should raise `tanjun.errors.HaltExecution` to
         end the execution search when it fails instead of returning `False`.
 
         Defaults to `False`.
-    owner_ids: typing.Optional[collections.abc.Iterable[hikari.snowflakes.SonwflakeishOr[hikari.users.User]]]
-        Iterable of objects and IDs of other users to explicitly mark as owners
-        for this check.
 
     Notes
     -----
-    * Any provided `owner_ids` will be used alongside the application's owners.
-    * This is based on the owner(s) of the bot's application and will account
-      for team owners as well.
     * error_message takes priority over halt_execution.
     * For more information on how this is used with other parameters see
       `CallbackReturnT`.
@@ -699,12 +594,7 @@ def with_owner_check(
     CallbackReturnT[CommandT]
         The command this check was added to.
     """
-    return _wrap_with_kwargs(
-        command,
-        ApplicationOwnerCheck(
-            halt_execution=halt_execution, error_message=error_message, expire_delta=expire_delta, owner_ids=owner_ids
-        ),
-    )
+    return _optional_kwargs(command, OwnerCheck(halt_execution=halt_execution, error_message=error_message))
 
 
 def with_author_permission_check(